+++ R/FilterStatesMAE.R
@@ -4,6 +4,8 @@
 MAEFilterStates <- R6::R6Class( # nolint
   classname = "MAEFilterStates",
   inherit = FilterStates,
+
+  # public methods ----
   public = list(
     #' @description Initializes `MAEFilterStates` object
     #'
@@ -32,9 +34,9 @@ MAEFilterStates <- R6::R6Class( # nolint
       private$keys <- keys
       private$varlabels <- varlabels
 
-      self$queue_initialize(
+      self$state_list_initialize(
         list(
-          y = ReactiveQueue$new()
+          y = reactiveVal()
         )
       )
       return(invisible(self))
@@ -48,9 +50,11 @@ MAEFilterStates <- R6::R6Class( # nolint
     format = function(indent = 0) {
       checkmate::assert_number(indent, finite = TRUE, lower = 0)
 
-      if (length(self$queue_get(1L)) > 0) {
+      if (length(self$state_list_get(1L)) > 0) {
         formatted_states <- sprintf("%sSubject filters:", format("", width = indent))
-        for (state in self$queue_get(1L)) formatted_states <- c(formatted_states, state$format(indent = indent + 2))
+        for (state in self$state_list_get(1L)) {
+          formatted_states <- c(formatted_states, state$format(indent = indent + 2))
+        }
         paste(formatted_states, collapse = "\n")
       }
     },
@@ -72,26 +76,26 @@ MAEFilterStates <- R6::R6Class( # nolint
       moduleServer(
         id = id,
         function(input, output, session) {
-          previous_state <- reactiveVal(isolate(self$queue_get("y")))
+          previous_state <- reactiveVal(isolate(self$state_list_get("y")))
           added_state_name <- reactiveVal(character(0))
           removed_state_name <- reactiveVal(character(0))
 
-          observeEvent(self$queue_get("y"), {
-            added_state_name(setdiff(names(self$queue_get("y")), names(previous_state())))
-            removed_state_name(setdiff(names(previous_state()), names(self$queue_get("y"))))
+          observeEvent(self$state_list_get("y"), {
+            added_state_name(setdiff(names(self$state_list_get("y")), names(previous_state())))
+            removed_state_name(setdiff(names(previous_state()), names(self$state_list_get("y"))))
 
-            previous_state(self$queue_get("y"))
+            previous_state(self$state_list_get("y"))
           })
 
           observeEvent(added_state_name(), ignoreNULL = TRUE, {
-            fstates <- self$queue_get("y")
+            fstates <- self$state_list_get("y")
             html_ids <- private$map_vars_to_html_ids(names(fstates))
             for (fname in added_state_name()) {
               private$insert_filter_state_ui(
                 id = html_ids[fname],
                 filter_state = fstates[[fname]],
-                queue_index = "y",
-                element_id = fname
+                state_list_index = "y",
+                state_id = fname
               )
             }
             added_state_name(character(0))
@@ -117,7 +121,7 @@ MAEFilterStates <- R6::R6Class( # nolint
     #'
     #' @return `list` with elements number equal number of `FilterStates`.
     get_filter_state = function() {
-      lapply(self$queue_get(queue_index = "y"), function(x) x$get_state())
+      lapply(self$state_list_get(state_list_index = "y"), function(x) x$get_state())
     },
 
     #' @description
@@ -139,7 +143,7 @@ MAEFilterStates <- R6::R6Class( # nolint
         combine = "or"
       )
       logger::log_trace("MAEFilterState$set_filter_state initializing, dataname: { deparse1(private$input_dataname) }")
-      filter_states <- self$queue_get("y")
+      filter_states <- self$state_list_get("y")
       for (varname in names(state)) {
         value <- resolve_state(state[[varname]])
         if (varname %in% names(filter_states)) {
@@ -155,10 +159,10 @@ MAEFilterStates <- R6::R6Class( # nolint
           )
           fstate$set_state(value)
           fstate$set_na_rm(TRUE)
-          self$queue_push(
+          self$state_list_push(
             x = fstate,
-            queue_index = "y",
-            element_id = varname
+            state_list_index = "y",
+            state_id = varname
           )
         }
       }
@@ -167,46 +171,48 @@ MAEFilterStates <- R6::R6Class( # nolint
     },
 
     #' @description
-    #' Removes a variable from the `ReactiveQueue` and its corresponding UI element.
+    #' Removes a variable from the `state_list` and its corresponding UI element.
     #'
-    #' @param element_id (`character(1)`)\cr name of `ReactiveQueue` element.
+    #' @param state_id (`character(1)`)\cr name of `state_list` element.
     #'
     #' @return `NULL`
     #'
-    remove_filter_state = function(element_id) {
+    remove_filter_state = function(state_id) {
       logger::log_trace(
         sprintf(
           "%s$remove_filter_state for %s called, dataname: %s",
           class(self)[1],
-          element_id,
+          state_id,
           deparse1(private$input_dataname)
         )
       )
 
-      if (!element_id %in% names(self$queue_get("y"))) {
+      if (!state_id %in% names(self$state_list_get("y"))) {
         warning(paste(
-          "Variable:", element_id,
+          "Variable:", state_id,
           "is not present in the actual active filters of dataset: { deparse1(private$input_dataname) }",
           "therefore no changes are applied."
         ))
         logger::log_warn(
           paste(
-            "Variable:", element_id, "is not present in the actual active filters of dataset:",
+            "Variable:", state_id, "is not present in the actual active filters of dataset:",
             "{ deparse1(private$input_dataname) } therefore no changes are applied."
           )
         )
       } else {
-        self$queue_remove(queue_index = "y", element_id = element_id)
+        self$state_list_remove(state_list_index = "y", state_id = state_id)
         logger::log_trace(
           sprintf(
             "%s$remove_filter_state for variable %s done, dataname: %s",
             class(self)[1],
-            element_id,
+            state_id,
             deparse1(private$input_dataname)
           )
         )
       }
     },
+
+    # shiny modules ----
 
     #' @description
     #' Shiny UI module to add filter variable
@@ -214,8 +220,10 @@ MAEFilterStates <- R6::R6Class( # nolint
     #'  id of shiny module
     #' @param data (`MultiAssayExperiment`)\cr
     #'  object containing `colData` which columns are used to be used
-    #'  to choose filter variables.
-    #' @return shiny.tag
+    #'  to choose filter variables
+    #'
+    #' @return `shiny.tag`
+    #'
     ui_add_filter_state = function(id, data) {
       checkmate::assert_string(id)
       stopifnot(is(data, "MultiAssayExperiment"))
@@ -251,7 +259,9 @@ MAEFilterStates <- R6::R6Class( # nolint
     #'  object containing `colData` which columns are used to choose filter variables in
     #' [teal.widgets::optionalSelectInput()].
     #' @param ... ignored
+    #'
     #' @return `moduleServer` function which returns `NULL`
+    #'
     srv_add_filter_state = function(id, data, ...) {
       stopifnot(is(data, "MultiAssayExperiment"))
       check_ellipsis(..., stop = FALSE)
@@ -264,7 +274,7 @@ MAEFilterStates <- R6::R6Class( # nolint
           shiny::setBookmarkExclude("var_to_add")
           active_filter_vars <- reactive({
             vapply(
-              X = self$queue_get(queue_index = "y"),
+              X = self$state_list_get(state_list_index = "y"),
               FUN.VALUE = character(1),
               FUN = function(x) x$get_varname(deparse = TRUE)
             )
@@ -328,10 +338,10 @@ MAEFilterStates <- R6::R6Class( # nolint
               )
               fstate$set_na_rm(TRUE)
 
-              self$queue_push(
+              self$state_list_push(
                 x = fstate,
-                queue_index = "y",
-                element_id = input$var_to_add
+                state_list_index = "y",
+                state_id = input$var_to_add
               )
               logger::log_trace(
                 sprintf(
@@ -351,6 +361,8 @@ MAEFilterStates <- R6::R6Class( # nolint
       )
     }
   ),
+
+  # private members ----
   private = list(
     varlabels = character(0),
     keys = character(0),