+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Examples now use `scda.2022` instead of `scda.2021`.
 * Transferred data hashing step in `FilteredDataset` to `teal`.
+* Removed constructor of `Queue` class.
 
 ### New features